+++ app/Models/Text.php
@@ -24,6 +24,7 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use App\Contracts\HasComments;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\MorphMany;
 use Illuminate\Database\Eloquent\SoftDeletes;
@@ -40,7 +41,7 @@ use Spatie\Translatable\HasTranslations;
  * @property Source|null $originText
  * @property Source|null $copyrightText
  */
-class Text extends Model
+class Text extends Model implements HasComments
 {
     use HasFactory, HasTranslations, LogsActivity,SoftDeletes;
 
@@ -88,7 +89,7 @@ class Text extends Model
      *
      * @return MorphMany
      */
-    public function comments()
+    public function comments(): MorphMany
     {
         return $this->morphMany(Comment::class, 'commentable')->whereNull('parent_id');
     }